TASK: Move `NodeCreationHandler` implementations to fitting sub namespace
