+++ Classes/Infrastructure/ContentRepository/PromotedElementsCreationHandlerFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Neos\Neos\Ui\Infrastructure\NodeCreation;
+namespace Neos\Neos\Ui\Infrastructure\ContentRepository;
 
 use Neos\ContentRepository\Core\Factory\ContentRepositoryServiceFactoryDependencies;
 use Neos\ContentRepository\Core\Factory\ContentRepositoryServiceFactoryInterface;
