+++ Configuration/NodeTypes.yaml
@@ -19,15 +19,15 @@
   options:
     nodeCreationHandlers:
       uriPathSegment:
-        factoryClassName: 'Neos\Neos\Ui\Infrastructure\NodeCreation\UriPathSegmentNodeCreationHandlerFactory'
+        factoryClassName: 'Neos\Neos\Ui\Infrastructure\Neos\UriPathSegmentNodeCreationHandlerFactory'
       promotedElements:
-        factoryClassName: 'Neos\Neos\Ui\Infrastructure\NodeCreation\PromotedElementsCreationHandlerFactory'
+        factoryClassName: 'Neos\Neos\Ui\Infrastructure\ContentRepository\PromotedElementsCreationHandlerFactory'
 
 'Neos.Neos:Content':
   options:
     nodeCreationHandlers:
       promotedElements:
-        factoryClassName: 'Neos\Neos\Ui\Infrastructure\NodeCreation\PromotedElementsCreationHandlerFactory'
+        factoryClassName: 'Neos\Neos\Ui\Infrastructure\ContentRepository\PromotedElementsCreationHandlerFactory'
 
 'Neos.Neos:ContentCollection':
   ui:
